Move executable scripts to a separate folder
